// dpbloom-client/src/app/core/api/model/submitAnswerDto.ts
/**
 * DPBloom | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SubmitAnswerDto { 
    questionId?: string;
    selectedOptionIds?: Array<string> | null;
    freeTextAnswer?: string | null;
    id?: string;
    createdOn?: string | null;
    updatedOn?: string | null;
}